pyproject.toml
[project]
name = "drifter"
version = "0.1.0"
description = "Modular, production-grade RAG system built from first principles"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.66.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
adapters = [
    "qdrant-client>=1.9",
    "opensearch-py>=2.5",
    "psycopg[binary]>=3.1",
    "boto3>=1.34",
]
observability = [
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-http>=1.24",
]

[project.scripts]
rag = "apps.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["libs", "orchestrators", "apps"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
